Update for browser compatibility:
Compatibility:
- Add "scroll" polyfill for smooth scrolling

Other:
- Change Section.js to AboutMe.js
- Change Section.less to AboutMe.less
- Refactor code to include former Section.js contents to be within AboutMe.js
