--- js/content/Headshot.js
@@ -1,13 +1,0 @@
-import React from 'react'
-
-const Headshot = React.createClass({
-  render () {
-    return (
-      <div className='headshot'>
-        <img className='img-responsive' src='../../public/img/tim_headshot.png' />
-      </div>
-    )
-  }
-})
-
-export default Headshot
