+++ package.json
@@ -27,7 +27,8 @@
     "react-redux": "4.4.0",
     "react-router": "4.0.0-alpha.5",
     "redux": "3.3.1",
-    "redux-thunk": "^2.1.0"
+    "redux-thunk": "^2.1.0",
+    "smoothscroll-polyfill": "^0.3.4"
   },
   "devDependencies": {
     "babel-core": "6.16.0",
