+++ public/less/Landing.less
@@ -9,7 +9,7 @@
     position:fixed;
     height: 100vh;
     z-index: -1;
-    filter: sepia(50%;);
+    filter: sepia(60%;);
   }
   .section-break {
     display: block;
