+++ public/less/Value.less
@@ -18,6 +18,7 @@
     img{
       max-height: 300px;
       margin:auto;
+      opacity: .5!important;
     }
   }
   .value-content {
