+++ public/less/common.less
@@ -4,7 +4,6 @@
 @import 'Landing.less';
 @import 'Navbar.less';
 @import 'Jumbotron.less';
-@import 'Section.less';
 @import 'AboutMe.less';
 @import 'Social.less';
 @import 'MyValues.less';
